project1/pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "streamlit>=1.55.0",
    "tensorflow>=2.20.0",
]
